python/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
flask-restx==0.5.1
requests==2.32.3
PyNaCl==1.5.0
cryptography==42.0.8
pycryptodome==3.20.0
